% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banquo.R
\name{banquo}
\alias{Banquo}
\alias{banquo}
\title{Turn `enquo` into `banquo`}
\usage{
banquo(arg)

Banquo(arg)
}
\arguments{
\item{expr}{An expression.}
}
\value{
A formula whose right-hand side contains the quoted
  expression supplied as argument.
}
\description{
Did you also see \code{rlang}'s \code{\link[rlang]{enquo}} function and instantly think of your favorite
Thane of Lochaber, Shakespeare's very own Banquo?

No?

Well, this function lets you call \code{enquo()} with \code{Banquo()} (or \code{banquo()}). That's it.
}
\examples{
yell_at_son <- function(fleance_cry) {
  fleance_cry <- Banquo(fleance_cry)
  tell_fleance <- paste0(rlang::quo_name(fleance_cry),
                         ", good Fleance, ",
                         rlang::quo_name(fleance_cry),
                         ", ",
                         rlang::quo_name(fleance_cry),
                         ", ",
                         rlang::quo_name(fleance_cry),
                         "!")
 print(paste0(toupper(substr(tell_fleance, 1, 1)),
              substr(tell_fleance, 2, nchar(tell_fleance))))
}
yell_at_son(fly)
yell_at_son(cry)
}
\seealso{
See \code{\link[rlang]{enquo}} for details
}

